Backport patch that removes 'Page 1/1' counter
I submitted this change upstream to pelican, but it won't make it back into our
theme. Hence, I'm backporting it here.

When you have only one page of articles in a category (e.g. any of the tags
with one item), you'll find that "Page 1/1" is redundantly displayed at the
bottom. This fix prevents that message from being displayed.